Added cloaking
Wraiths and ghosts now cloak "as they're supposed to".
This feature is still a bit shaky, and not tested completely.
Most of the shakiness comes from units going in and out of cloak a lot.
This is because detectors and enemy units often go in and out of sight.
I also had to change from Physics Lab to Covert Ops in the
StructureAgent.
Debug printouts are intentionally left, for further testing.