Implemented Lockdown
Ghosts can now use the Lockdown ability.
Also added a predicate that returns a collection of robotic AND
mechanical units, solely for the purpose of using Lockdown.
This feature builds properly but is untested, therefore not added
to master branch. It is up to date, however and should work.
A Ghost will try to use lockdown iff:
- The enemy is within sight range
- The enemy is mechanical or robotic
- The enemy isn't already locked down

Additionally, if there are several units matching the criteria,
the Ghost will target the one with the highest destroy cost.